refactor: adicionando hooks para use states que ficam grandes
